requirements.txt
pandas
numpy
sentence-transformers
umap-learn
